# Conflicts:
#	src/locales/de.json
#	src/locales/en.json

+++ src/backend/admin/features/adminPermission/DownloadUpdate.php
@@ -10,21 +10,24 @@ use backend\exceptions\PageFlowException;
 class DownloadUpdate extends HasAdminPermission {
 	
 	function exec(): array {
-		$preRelease = (bool) $_GET['preRelease'];
-		$versionString = $_GET['version'];
+		if(!isset($_POST['url']))
+			throw new PageFlowException('Missing data');
+		
+		$url = $_POST['url'];
 		
 		$pathUpdate = Paths::FILE_SERVER_UPDATE;
 		if(file_exists($pathUpdate))
 			unlink($pathUpdate);
 		
-		$url = sprintf(Configs::get($preRelease ? 'url_update_preReleaseZip' : 'url_update_releaseZip'), $versionString);
-		
 		$res = @fopen($url, 'r');
 		if(!$res)
 			throw new PageFlowException("Downloading update from $url failed. Nothing was changed");
 		
-		if(!@file_put_contents($pathUpdate, $res))
+		if(!@file_put_contents($pathUpdate, $res)) {
+			fclose($res);
 			throw new PageFlowException("Saving update file to $pathUpdate failed. Nothing was changed");
+		}
+		fclose($res);
 		
 		return [];
 	}
